kernel: git: restore submodules to fix recurse cloning
This fixes errors when this kernel repository is cloned with
--recurse-submodules, due to containing other git repos nested within,
but without .gitmodules file. The error git clone fails with is:

fatal: No url found for submodule path
'kernel/sony/msm-4.9/kernel/drivers/staging/wlan-qc/fw-api' in .gitmodules

This is essential for use cases where this repository is a submodule itself,
and recursive cloning is the only policy used during the CI builds.

[kernel] git: restore submodules to fix recurse cloning. JB#51476

Signed-off-by: Simonas Leleiva <simonas.leleiva@jolla.com>